==> 0_preprocessing.ipynb <==
pip install pandas

import pandas as pd
import warnings

warnings.filterwarnings('ignore')

df0 = pd.read_json("data/Streaming_History_Audio_2019-2020_0.json")
df1 = pd.read_json("data/Streaming_History_Audio_2020-2022_1.json")
df2 = pd.read_json("data/Streaming_History_Audio_2022-2024_2.json")
df3 = pd.read_json("data/Streaming_History_Audio_2024_3.json")

for df in [df0, df1, df2, df3]:
  print(df['ts'][0], df['ts'][len(df)-1])

df_all = pd.concat([df0, df1, df2, df3]).reset_index(drop=True)
print(len(df_all))
df_all

df = df_all[['ts', 'ms_played', 'master_metadata_track_name', 'master_metadata_album_artist_name','master_metadata_album_album_name']]
df.head(2)

df.isna().sum()

df = df.dropna()

df.dtypes

tmp = pd.to_datetime(df['ts']).dt.strftime('%Y-%m-%d %H-%M-%S')
df['ts'] = pd.to_datetime(tmp, format='%Y-%m-%d %H-%M-%S')
df.head(5)

df['ms_played'] = round(df['ms_played'] / 60000, 2)
df['ms_played'].head(1)

Shorten song and album names by eliminating parentheses and brackets. For example,

import re

def shorten_string(text):
  # Remove substrings inside parentheses or brackets
  text = re.sub(r"\(.*?\)|\[.*?\]", "", text)
  # Trim the string
  text = text.strip()
  return text

df['master_metadata_track_name'] = df['master_metadata_track_name'].apply(shorten_string)
df['master_metadata_track_name'] = df['master_metadata_track_name'].replace('Nuvole bianche', 'Nuvole Bianche')
df.head(5)

df['master_metadata_album_album_name'] = df['master_metadata_album_album_name'].apply(shorten_string)
df.head(5)

df['song_artist'] = df['master_metadata_track_name'] + ' - ' + df['master_metadata_album_artist_name']
df.head(1)

Rename Columns

df.columns = ['timestamp', 'duration','song','artist','album','song_artist']
df.head(1)

df.to_csv("0_preprocessed.csv")
