==> 1_bar_chart_race.ipynb <==
pip install pandas matplotlib bar_chart_race

import pandas as pd
import warnings

warnings.filterwarnings('ignore')

df = pd.read_csv("data/0_preprocessed.csv")
df.head(5)

df.isna().sum()

df = df.dropna()

df.dtypes

tmp = pd.to_datetime(df['timestamp']).dt.strftime('%Y-%m-%d %H-%M-%S')
df['timestamp'] = pd.to_datetime(tmp, format='%Y-%m-%d %H-%M-%S')
df.head(5)

df.head(2)

from datetime import datetime, timedelta

# Create a list to store the datetime objects
timestamps = []

start_date = datetime(2019, 7, 5)
end_date = datetime(2024, 9, 6)

num_weeks = (end_date - start_date).days // 7

for i in range(num_weeks + 1):
  timestamps.append(start_date + timedelta(weeks=i))

# Print the list of datetime objects
print(len(timestamps), timestamps[0], timestamps[-1])

top_tracks_list = df['song_artist'].unique().tolist()

duration_dict_song = {song: [0 for _ in range(len(timestamps))] for song in top_tracks_list}
duration_total_dict_song = {song: 0 for song in top_tracks_list}

print(df['timestamp'][0], timestamps[0], df['timestamp'][0] < timestamps[0])

time_idx = 1
for i, row in df.iterrows():
  while row['timestamp'] > timestamps[time_idx]:
    for s in top_tracks_list:
      duration_dict_song[s][time_idx] = duration_total_dict_song[s]
    time_idx += 1

  song_artist = row['song_artist']
  duration_total_dict_song[song_artist] += row['duration']
  duration_dict_song[song_artist][time_idx] = duration_total_dict_song[song_artist]

df_pivot_song = pd.DataFrame(duration_dict_song, index=timestamps)
print(len(df_pivot_song), len(df_pivot_song.columns))
df_pivot_song

# Drop last row
df_pivot_song = df_pivot_song[:-1]
df_pivot_song.shape, df_pivot_song.index[0], df_pivot_song.index[-1]

artists = df['artist'].unique().tolist()
print(len(artists))

duration_dict_artist = {song: [0 for _ in range(len(timestamps))] for song in artists}
duration_total_dict_artist = {song: 0 for song in artists}

time_idx = 1
for i, row in df.iterrows():
  while row['timestamp'] > timestamps[time_idx]:
    for s in artists:
      duration_dict_artist[s][time_idx] = duration_total_dict_artist[s]
    time_idx += 1

  artist = row['artist']
  duration_total_dict_artist[artist] += row['duration']
  duration_dict_artist[artist][time_idx] = duration_total_dict_artist[artist]

df_pivot_artist = pd.DataFrame(duration_dict_artist, index=timestamps)
print(len(df_pivot_artist), len(df_pivot_artist.columns))
df_pivot_artist

# Drop last row
df_pivot_artist = df_pivot_artist[:-1]
df_pivot_artist.shape, df_pivot_artist.index[0], df_pivot_artist.index[-1]

# Check if values in each column are increasing
for col in df_pivot_artist.columns:
  prev_val = 0
  for i, val in enumerate(df_pivot_artist[col]):
    if val < prev_val:
      print(col)
      break
    prev_val = val

import bar_chart_race as bcr

# Create the bar chart race
bcr.bar_chart_race(
    df=df_pivot_song,
    title='Top 10 Most-played Songs (min)',
    filename='spotify_top_10_songs.mp4',  # Save as a file
    bar_label_size=17,
    tick_label_size=10,
    cmap='dark12',  # Color map for the bars
    figsize=(12, 6),
    period_length=400,  # Duration of each frame
    filter_column_colors=True,
    period_fmt='%b %-d, %Y',
    n_bars=10  # Limit to top 10 songs
)

# from IPython.display import Video

# # Display the video
# Video('scratch.mp4', embed=True)

# Create the bar chart race
bcr.bar_chart_race(
    df=df_pivot_artist,
    title='Top 10 Most-played Artists (min)',
    filename='spotify_top_10_artists.mp4',  # Save as a file
    bar_label_size=17,
    tick_label_size=10,
    cmap='dark12',  # Color map for the bars
    figsize=(12, 6),
    period_length=400,  # Duration of each frame
    filter_column_colors=True,
    period_fmt='%b %-d, %Y',
    n_bars=10  # Limit to top 10 songs
)

# pip install moviepy
# from moviepy.editor import VideoFileClip, concatenate_videoclips

# # Load the original video
# original_clip = VideoFileClip('final.mp4')

# # Extract the last frame
# last_frame = original_clip.get_frame(original_clip.duration)

# # Create a clip from the last frame
# from moviepy.editor import ImageClip
# last_frame_clip = ImageClip(last_frame).set_duration(5)  # Duration in seconds

# # Set the fps of the last frame clip to match the original video
# last_frame_clip = last_frame_clip.set_fps(original_clip.fps)

# # Concatenate the original video with the last frame clip
# extended_clip = concatenate_videoclips([original_clip, last_frame_clip])

# # Write the result to a new file
# extended_clip.write_videofile('top_10_extended.mp4')


Extract the data

df_pivot_artist.to_csv('df_pivot_artist.csv')
df_pivot_song.to_csv('df_pivot_song.csv')
