==> 2_sankey_top_100.ipynb <==
! pip install pandas plotly

import pandas as pd
import warnings

warnings.filterwarnings('ignore')

df0 = pd.read_csv('data/0_preprocessed.csv')
df0

top_100 = df0.groupby(['artist', 'song'])['duration'].sum().reset_index()
top_100 = top_100.sort_values('duration', ascending=False).head(100).reset_index(drop=True)
top_100['duration'] = top_100['duration'].astype(int)
top_100

len(top_100['artist'].unique())

top_100.to_csv('2_sankey_top_100.csv', index=False)

Manually add columnn `type` and `genre`

df = pd.read_csv('2_sankey_top_100.csv')
df

Manually fill in type and genres

import pandas as pd

df2 = pd.read_csv('2_sankey_top_100.csv')
df2

len(df2['type'].unique()), len(df2['genre'].unique()), len(df2['artist'].unique())

labels = ['Most-played 100']
labels += df2['type'].unique().tolist() + df2['genre'].unique().tolist() + df2['artist'].unique().tolist()
labels_dict = {label: i for i, label in enumerate(labels)}
len(labels), len(labels_dict)

links = {}
for i, row in df2.iterrows():
  tmp = ['Most-played 100', row['type'], row['genre'], row['artist']]
  for source, target in zip(tmp, tmp[1:]):
    link = f"{labels_dict[source]}-{labels_dict[target]}"
    if link not in links:
      links[link] = 0
    links[link] += 1

len(links)

sources, targets, values = [], [], []
for link, value in links.items():
  source, target = link.split('-')
  sources.append(int(source))
  targets.append(int(target))
  values.append(value)

len(sources), len(targets), len(values)

import json
data = {
    "labels": labels,
    "sources": sources,
    "targets": targets,
    "values": values
}

data_json = json.dumps(data, indent=4)

with open('2_sankey_top_100.json', 'w') as f:
  f.write(data_json)

import plotly.graph_objects as go

fig = go.Figure(data=[go.Sankey(
    node = dict(
      pad = 15,
      thickness = 20,
      line = dict(color = "black", width = 0.5),
      label = labels,
      hoverinfo = None
      # color = "blue"
    ),
    link = dict(
      source = sources,
      target = targets,
      value = values,
  ))])

fig.update_layout(title_text="Basic Sankey Diagram", font_size=10)
fig.show()

?go.Figure
