==> 4_sunburst_albums.ipynb <==
!pip install pandas plotly numpy

import pandas as pd
import numpy as np

df0 = pd.read_csv('data/0_preprocessed.csv')
df0.head(2)

# Get albums with 5 or more distinct songs
albums = df0.groupby(['album', 'artist'])['song'].nunique().reset_index()
albums = albums[albums['song'] >= 5]

#Get album duration
album_duration = df0.groupby(['album', 'artist'])['duration'].sum().reset_index()

df4_albums = pd.merge(albums, album_duration, on=['album', 'artist'])
df4_albums = df4_albums.rename(columns={
    'album': 'child',
    'artist': 'parent',
    'song': 'n_songs'})
df4_albums['duration'] = df4_albums['duration'].astype(int)
df4_albums = df4_albums.sort_values('duration', ascending=False)
df4_albums = df4_albums[df4_albums['child'] != df4_albums['parent']]
df4_albums = df4_albums.reset_index(drop=True)
df4_albums.head(10)

fearless = df4_albums[df4_albums['child'].str.contains('Fearless')]
df4_albums = df4_albums[~df4_albums['child'].str.contains('Fearless')]

new_row = pd.DataFrame({
    'child': ['Fearless'],
    'parent': ['Taylor Swift'],
    'n_songs': [fearless.n_songs.sum()],
    'duration': [fearless.duration.sum()]
})
df4_albums = pd.concat([df4_albums, new_row]).sort_values('duration', ascending=False).reset_index(drop=True)
df4_albums

df4_artists = df4_albums.groupby('parent').agg({'n_songs': 'sum', 'duration': 'sum'}).reset_index()
df4_artists.rename(columns={'parent': 'child'}, inplace=True)
df4_artists.insert(2, 'parent', 'Top Albums')
new_row = pd.DataFrame({
    'parent': [''],
    'child': ['Top Albums'],
    'n_songs': [df4_artists.n_songs.sum()],
    'duration': [int(df4_artists.duration.sum())]
})
df4_artists = pd.concat([new_row, df4_artists], ignore_index=True)
df4_artists

df4 = pd.concat([df4_artists, df4_albums], ignore_index=True)
df4['scale'] = np.log10(df4['duration']).astype(int)
df4.loc[0, 'scale'] = 0
df4

import plotly.express as px

fig = px.sunburst(
    df4,
    names='child',
    parents='parent',
    values='scale',
    hover_data={
        'child': False,
        'parent': False,
        'n_songs': True,
        'duration': True,
        'scale': False
    },
)
fig.show()

df4.to_csv('4_sunburst_albums.csv', index=False)
