==> 5_taylor_top_10.ipynb <==
pip install pandas

import pandas as pd

df = pd.read_csv('data/0_preprocessed.csv')
df

taylor_top_10_list = [
 'Sparks Fly',
 'Enchanted',
 'Begin Again',
 'If This Was A Movie',
 'Treacherous',
 'I\'d Lie',
 'I Almost Do',
 'Everything Has Changed',
 'Fearless',
 'Superman'
]

df_taylor = df[df['artist'] == 'Taylor Swift']
df_taylor = df_taylor.groupby(['song', 'album'])['duration'].sum().reset_index().sort_values(by='duration', ascending=False)
df_taylor['duration'] = df_taylor['duration'].astype(int)
df_taylor

taylor_top_10 = pd.DataFrame({
  'n': [i for i in range(1,11)],
  'most_played': df_taylor.head(10)['song'].tolist(),
  'mins': df_taylor.head(10)['duration'].tolist(),
  'fav_10': taylor_top_10_list
})

taylor_top_10

my_fav_10_ranked = [50,16,42,218,63,207,117,130,41,220]
rs_top_10 = [
  "All Too Well",
  "New Romantics",
  "mirrorball",
  "Delicate",
  "august",
  "Long Live",
  "Lover",
  "Blank Space",
  "majorie",
  "Dear John"
]

taylor_top_10['RS_ranked'] = my_fav_10_ranked
taylor_top_10['RS_top_10'] = rs_top_10
taylor_top_10

taylor_top_10.to_csv('5_taylor_top_10.csv', index=False)
