==> 6_diverging_bar_chart.ipynb <==
!pip install pandas plotly

import pandas as pd

df0 = pd.read_csv('data/0_preprocessed.csv')
df0

df0 = df0[(df0['artist'] == 'Westlife') | (df0['artist'] == 'Backstreet Boys')]
df0 = df0.groupby(['song', 'album', 'artist'])['duration'].sum().astype(int).reset_index()
df0 = df0.sort_values('duration', ascending=False).reset_index(drop=True)
df0

bsb = df0[df0['artist'] == 'Backstreet Boys'][["song", "duration"]].head(5).reset_index(drop=True)
wl = df0[df0['artist'] == 'Westlife'][["song", "duration"]].head(5).reset_index(drop=True)

df6 = pd.concat([bsb, wl], axis=1)
df6.columns = ['bsb','bsb_duration','westlife','westlife_duration']
df6

new_row = [{
  'bsb': 'Backstreet Boys',
  'bsb_duration': df0[df0['artist'] == 'Backstreet Boys']['duration'].sum(),
  'westlife': 'Westlife',
  'westlife_duration': df0[df0['artist'] == 'Westlife']['duration'].sum()
}]

df6 = pd.concat([pd.DataFrame(new_row), df6], ignore_index=True)
df6

total_duration = df6['bsb_duration'] + df6['westlife_duration']
df6['bsb_percentage'] = round(((df6['bsb_duration'] / total_duration) * 100), 0).astype(int)
df6['westlife_percentage'] = round(((df6['westlife_duration'] / total_duration) * 100), 0).astype(int)
df6

df6.to_csv('6_diverging_bar_chart.csv', index=False)
