==> 7_disney_treemap.ipynb <==
!pip install pandas plotly

import pandas as pd

df0 = pd.read_csv('data/0_preprocessed.csv')
df0

disney_songs = [
    "For the First Time in Forever - From \"Frozen\"/Soundtrack Version",
    "Let It Go - From \"Frozen\"/Soundtrack Version",
    "Love Is an Open Door - From \"Frozen\"/Soundtrack Version",
    "Into the Unknown",
    "Show Yourself",
    "Some Things Never Change - From \"Frozen 2\"/Soundtrack Version",
    "All Is Found",
    "Speechless",
    "A Whole New World",
    "I See the Light - From \"Tangled\" / Soundtrack Version",
    "When Will My Life Begin? - From \"Tangled\" / Soundtrack Version",
    "How Far I'll Go - From \"Moana\"/Soundtrack Version"
]

df_disney = df0[df0['song'].isin(disney_songs)]
len(df_disney), df_disney.song.unique()

disney_songs_renamed = [
    "For the First Time in Forever",
    "Let It Go",
    "Love Is an Open Door",
    "Into the Unknown",
    "Show Yourself",
    "Some Things Never Change",
    "All Is Found",
    "Speechless",
    "A Whole New World",
    "I See the Light",
    "When Will My Life Begin?",
    "How Far I'll Go"
]

disney_movies = {
    "For the First Time in Forever": "Frozen",
    "Let It Go": "Frozen",
    "Love Is an Open Door": "Frozen",
    "Into the Unknown": "Frozen 2",
    "Show Yourself": "Frozen 2",
    "Some Things Never Change": "Frozen 2",
    "All Is Found": "Frozen 2",
    "Speechless": "Aladdin",
    "A Whole New World": "Aladdin",
    "I See the Light": "Tangled",
    "When Will My Life Begin?": "Tangled",
    "How Far I'll Go": "Moana"
}

df_disney = df_disney.groupby(['song'])['duration'].sum().reset_index().sort_values(by='duration', ascending=False).reset_index(drop=True)
df_disney['duration'] = df_disney['duration'].astype(int)
df_disney.reset_index(drop=True, inplace=True)
df_disney['song'] = df_disney['song'].map({disney_song: renamed for disney_song, renamed in zip(disney_songs, disney_songs_renamed)})
df_disney['movie'] = df_disney['song'].map(disney_movies)
# df_disney.loc[len(df_disney)] = ['Disney Music', df_disney.duration.sum(), 'All']
df_disney

import plotly.express as px
import numpy

fig = px.treemap(
    df_disney,
    path=[px.Constant("all"),'movie', 'song'],
    values='duration',
    color='duration',
    # hover_data=['iso_alpha'],
    color_continuous_scale='RdBu',
)
                  # color_continuous_midpoint=np.average(df['lifeExp'], weights=df['pop']))
fig.update_layout(margin = dict(t=50, l=25, r=25, b=25))
fig.show()


df_disney.to_csv('7_disney_treemap.csv', index=False)
