==> 8_waterfall_month.ipynb <==
pip install pandas

import pandas as pd
import warnings

warnings.filterwarnings('ignore')

df0 = pd.read_csv('data/0_preprocessed.csv')
df0

from datetime import datetime, timedelta

# Define the start and end dates
start_date = datetime(2019, 7, 1)
end_date = datetime(2024, 10, 1)

# Initialize an empty list to store the timestamps
timestamps = []

# Loop through each month from start_date to end_date
current_date = start_date
while current_date <= end_date:
    # Append the current date to the list
    timestamps.append(current_date)
    # Move to the first day of the next month
    next_month = current_date.month % 12 + 1
    next_year = current_date.year + (current_date.month // 12)
    current_date = datetime(next_year, next_month, 1)

len(timestamps), timestamps[0], timestamps[-1]

waterfall_duration = [] # Sum of activity every 7 days
waterfall_x = [] # Timestamp
waterfall_y = [] # Change of activity every 7 days
waterfall_text = [] # Change of activity as string text
activity = 0

for i in range(len(timestamps)-1):
  start_month = timestamps[i]
  end_month = timestamps[i+1]

  duration = df0[(df0['timestamp'] >= str(start_month)) & (df0['timestamp'] <= str(end_month))]['duration'].sum()
  duration = int(duration)
  y = duration - activity
  text = '+' + str(y) if y >= 0 else str(y)

  waterfall_duration.append(duration)
  waterfall_x.append(start_month)
  waterfall_y.append(y)
  waterfall_text.append(text)

  activity = duration

waterfall_n_weeks = [1]
waterfall_measure = ['total']

for i, timestamp in enumerate(waterfall_x[1:-1]):
  waterfall_n_weeks.append(i+1)
  if timestamp.month == 1:
    waterfall_measure.append('total')
  else:
    waterfall_measure.append('relative')

waterfall_n_weeks.append(len(waterfall_duration))
waterfall_measure.append("total")

len(waterfall_n_weeks), len(waterfall_measure)

waterfall_df = pd.DataFrame({
    'duration': waterfall_duration,
    'x': waterfall_x,
    'y': waterfall_y,
    'text': waterfall_text,
    'n_weeks': waterfall_n_weeks,
    'measure': waterfall_measure
})

waterfall_df

waterfall_df['measure'].value_counts()

waterfall_df['duration'].sum(), df0['duration'].sum()

waterfall_df.to_csv('8_waterfall.csv', index=False)
