==> 9_wind_rose_replayer.ipynb <==
!pip install pandas plotly

import pandas as pd

df0 = pd.read_csv('data/0_preprocessed.csv')
df0

cur_song, cur_duration = '', 0
replayed = {} # key: song, value: [a,b,c] a 10+ b 30+ c 60+

for i, row in df0.iterrows():
  if cur_song != row['song']:
    a,b,c = replayed.get(cur_song, [0,0,0])
    if cur_duration < 30:
      pass
    else:
      if cur_duration < 45:
        a += 1
      elif cur_duration < 60:
        b += 1
      else:
        c += 1
      replayed[cur_song] = [a,b,c]
    cur_duration = 0

  cur_song = row['song']
  cur_duration += row['duration']

len(replayed)

df9_js = pd.DataFrame.from_dict(replayed, orient='index', columns=['mins_30_45', 'mins_45_60', 'mins_60_plus'])
df9_js = df9_js.reset_index().rename(columns={'index': 'song'})
df9_js

def scale(x):
  if x == 0:
    x = 0
  elif 1 <= x <= 2:
    x = 5
  elif 3 <= x <= 5:
    x = 6
  elif 6 <= x <= 10:
    x = 7
  elif 11 <= x <= 15:
    x = 8
  else:
    x = 9
  return x

df9_js['mins_30_45'] = df9_js['mins_30_45'].apply(scale)
df9_js['mins_45_60'] = df9_js['mins_45_60'].apply(scale)
df9_js['mins_60_plus'] = df9_js['mins_60_plus'].apply(scale)
df9_js.head(2)

df9_js.to_csv('9_wind_rose_replayer_js.csv', index=False)

df9 = pd.DataFrame(replayed, index=['30-45 mins','45-60 mins','60+ mins']).transpose()
df9 = df9.reset_index().melt(id_vars='index', var_name='duration', value_name='times')
df9 = df9.rename(columns={'index': 'song'})
df9['scales'] = df9['times'].apply(scale)
df9

import plotly.express as px

hover_data = {
    'scales': False,
    'times': True,
    'duration': True,
    'song': True
}

fig = px.bar_polar(df9,
                   r="scales",
                   theta="song",
                   color="duration",
                   template="plotly_dark",
                  #  range_r=[0,20],
                  #  log_r=True,
                   color_discrete_sequence= px.colors.sequential.Plasma_r,
                   hover_data=hover_data,
                  #  showticklabels=False
                   )

fig.update_layout(
    polar=dict(radialaxis=dict(showticklabels=False, ticks='')),
    showlegend=False)
fig.show()

df9.to_csv('9_wind_rose_replayer.csv', index=False)

px.colors.sequential.Plasma_r
